bq-syncer/pyproject.toml
[tool.poetry]
name = "bq-syncer"
version = "0.1.0"
description = ""
authors = ["daoleno <daodaoonline@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
duckdb = "^0.9.2"
schedule = "^1.2.1"
google-cloud-bigquery = "^3.13.0"
polars = "^0.19.15"
pyarrow = "^14.0.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
